ChangesMethods
Cambios para hacer metodos de TransferBattery y TransferLoad, agregue a su vez metodos para calcular distancia y verificar si estan los operadores en la misma ubicacion. Cambie tambien el float de maxLoad por double para poder hacer calculos.

+++ SkyNet/SkyNet/K9.cs
@@ -15,7 +15,7 @@ namespace SkyNet
         public string SensorType { get; set; }
         public string Movility { get; set; }
       
-        public K9(/*string sensorType, string Movility, */string id, float maxLoad, Battery battery, Location location, string status) 
+        public K9(/*string sensorType, string Movility, */string id, double maxLoad, Battery battery, Location location, string status) 
             : base(maxLoad, battery, location, status, id)
         {
             //this.sensorType = string.Empty;
